refactor: move theme data to constants and centralize types
